mycelium/dashboard/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
asyncpg==0.29.0
nats-py==2.7.2
aiofiles==23.2.1
structlog==24.1.0
